Separates creation and building
Now we have a proper three-tiered system which can be layered in
more than just single inheritance.

Next feature is covering all the bases with a seed_fn, this comes
with a refactor that turns all of cluster into a truthy or nil
system with no internal throws.

Which we can wrap with an asserting closure quite easily.